+++ .eslintrc.js
@@ -1,14 +1,15 @@
 module.exports = {
-    parser: '@typescript-eslint/parser',
-    extends: ['plugin:@typescript-eslint/recommended', 'prettier/@typescript-eslint', 'plugin:prettier/recommended'],
+    parser: "@typescript-eslint/parser",
+    extends: ["plugin:@typescript-eslint/recommended", "prettier/@typescript-eslint", "plugin:prettier/recommended"],
     parserOptions: {
         ecmaVersion: 2019,
-        sourceType: 'module',
+        sourceType: "module",
     },
     rules: {
-        '@typescript-eslint/no-empty-function': 'off',
-        '@typescript-eslint/explicit-function-return-type': 'off',
-        '@typescript-eslint/ban-ts-ignore': 'off',
-        '@typescript-eslint/triple-slash-reference': 'off',
+        "@typescript-eslint/no-empty-function": "off",
+        "@typescript-eslint/explicit-function-return-type": "off",
+        "@typescript-eslint/ban-ts-ignore": "off",
+        "@typescript-eslint/triple-slash-reference": "off",
+        "@typescript-eslint/interface-name-prefix": "off",
     },
 };