+++ CHANGELOG.md
@@ -7,6 +7,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.0.6] - 2020-03-11
+
+### Added
+
+-   History indexedDB table [#8](https://github.com/Pageworks/wwibs/issues/8)
+-   Reply indexedDB table
+-   Reply table fallback for incognito users
+-   `reply()` method [#9](https://github.com/Pageworks/wwibs/issues/9)
+-   `replyAll()` method [#9](https://github.com/Pageworks/wwibs/issues/9)
+
+### Changed
+
+-   `broadcast-worker.ts` renamed to `wwibs-worker.ts`
+
+### Updated
+
+-   npm packages
+
 ## [0.0.5] - 2020-02-02
 
 ### Fixed
@@ -30,7 +48,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   prerelease script to prepare broadcaster worker script for CDN
 -   broadcaster export declarations
 
-[unreleased]: https://github.com/Pageworks/wwibs/compare/v0.0.5...HEAD
+[unreleased]: https://github.com/Pageworks/wwibs/compare/v0.0.6...HEAD
+[0.0.6]: https://github.com/Pageworks/wwibs/releases/tag/v0.0.6...v0.0.5
 [0.0.5]: https://github.com/Pageworks/wwibs/releases/tag/v0.0.5...v0.0.4
 [0.0.4]: https://github.com/Pageworks/wwibs/releases/tag/v0.0.4...v0.0.1
 [0.0.1]: https://github.com/Pageworks/wwibs/releases/tag/v0.0.1